src/requirements.txt
Flask
hvac
PyYAML
python-dateutil